Moved functions from wiring diagram scripts to Dec2022Demo.jl in ASKEM.jl. Made new augLabelledPetriNet function, accompanying functions, and incorporations into S2 wiring diagram. Made testS2.jl script in scripts to re-check stratification following changes.
